==> 4_word2vec_tracks.ipynb <==
# Tracks embbeding

import csv
import pandas as pd
import gensim
import os
import numpy as np
import sys
from tqdm import tqdm_notebook
import json
from gensim.models import Word2Vec
import time
from sklearn.manifold import TSNE
from matplotlib import pyplot as plt
from glob import glob
import random

mpd_sequence_path = 'mpd_playlist_track_seq/' #DEFINE YOUR PATH TO OUTPUT PLAYLIST TRACK ID sequences

sentences = [p.split() for p in open(mpd_sequence_path + "/tracks_sentences_id.txt").read().splitlines()]

# make model 
MODEL_SIZE = 300
EPOCHS = 1
WINDOW = 5 
MIN_COUNT = 5
lr = 50
NUM_FILES = 1000
name = "wv_model_mincount5_shuffle_size"+str(MODEL_SIZE)+"_MPD"

model = Word2Vec(sg=1, size=MODEL_SIZE, min_count=MIN_COUNT, window=WINDOW,compute_loss=True)
model.build_vocab(sentences)

if not os.path.exists("models"):
    os.makedirs("models")

# uncomment if you want to load a previous model 
"""
model = Word2Vec.load("models/wv_model_mincount5_shuffle_size300_MPD")
MODEL_SIZE = 300
EPOCHS = 1
WINDOW = 5 
MIN_COUNT = 5
lr = 70
NUM_FILES = 1000
name = "wv_model_mincount5_shuffle_size300_MPD"
"""

print("training..")
for iter in range(10000):# train
           
    model.train(sentences, total_examples=model.corpus_count, epochs=EPOCHS,compute_loss=True)
    [random.shuffle(l) for l in sentences]
               
    if(iter%10==0 and iter>0):
        # save model
        model.save("models/"+name)
        print("model "+name+" saved ") 
        
    print(model.get_latest_training_loss(),lr)
